% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMonitorMetricValue.R
\name{getMonitorMetricValue}
\alias{getMonitorMetricValue}
\title{任意の Metric 値を Azure Monitor API 経由にて取得}
\usage{
getMonitorMetricValue(SUBSCRIPTION_ID, RESOURCE_GROUP_NAME,
  RESOURCE_PROVIDER_NAMESPACE, RESOURCE_TYPE, RESOURCE_NAME, AZUREAD_TOKEN,
  AGGREGATION_TYPE, METRIC, TIMEGRAIN, STARTTIME, ENDTIME)
}
\arguments{
\item{SUBSCRIPTION_ID}{character}

\item{RESOURCE_GROUP_NAME}{character}

\item{RESOURCE_PROVIDER_NAMESPACE}{character}

\item{RESOURCE_TYPE}{character}

\item{RESOURCE_NAME}{character}

\item{AZUREAD_TOKEN}{character}

\item{AGGREGATION_TYPE}{character}

\item{METRIC}{character}

\item{TIMEGRAIN}{character}

\item{STARTTIME}{character}

\item{ENDTIME}{character}
}
\value{
result data.frame Metric の値
}
\description{
任意の Metric 値を Azure Monitor API 経由にて取得
}
