% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMonitorMetricValueDimensions.R
\name{getMonitorMetricValueDimensions}
\alias{getMonitorMetricValueDimensions}
\title{任意の Dimension Metric 値を Azure Monitor API 経由にて取得}
\usage{
getMonitorMetricValueDimensions(SUBSCRIPTION_ID, RESOURCE_GROUP_NAME,
  RESOURCE_PROVIDER_NAMESPACE, RESOURCE_TYPE, RESOURCE_NAME, AZUREAD_TOKEN,
  AGGREGATION_TYPE, METRIC, TIMEGRAIN, DIMENSIONS, STARTTIME, ENDTIME)
}
\arguments{
\item{SUBSCRIPTION_ID}{character}

\item{RESOURCE_GROUP_NAME}{character}

\item{RESOURCE_PROVIDER_NAMESPACE}{character}

\item{RESOURCE_TYPE}{character}

\item{RESOURCE_NAME}{character}

\item{AZUREAD_TOKEN}{character}

\item{AGGREGATION_TYPE}{character}

\item{METRIC}{character}

\item{TIMEGRAIN}{character}

\item{DIMENSIONS}{character}

\item{STARTTIME}{character}

\item{ENDTIME}{character}
}
\value{
result data.frame Dimension Metric の値
}
\description{
任意の Dimension Metric 値を Azure Monitor API 経由にて取得
}
